requirements.txt
asgiref==3.2.10
Django==2.1.5
djangorestframework==3.8.0
Pillow==8.0.0
pytz==2020.1
sqlparse==0.4.1
